# Conflicts:
#	README.md
#	docs/webserver-endpoints.md
#	lib/OpdsParser/OpdsParser.cpp
#	open-x4-sdk
#	src/CrossPointSettings.cpp
#	src/main.cpp
#	src/network/OtaUpdater.cpp

+++ src/activities/home/FileBrowserActivity.cpp
@@ -15,7 +15,7 @@
 #include "fontIds.h"
 
 namespace {
-constexpr unsigned long GO_HOME_MS = 1000;
+constexpr uint32_t GO_HOME_MS = 1000;
 }  // namespace
 
 void FileBrowserActivity::loadFiles() {
